backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-magic==0.4.27
dj-database-url==2.1.0
python-decouple==3.8 